Updates to Makefile and source files
The source files have been adjusted to directly refer to the link files needed for each
source file. New source files have been created to hold link directives required and the
Makefile file has been adjusted to refer to each file that now exists and the internal
requirements needed for compilation.

Testing of the new form has occurred and the results produced are identical to the previous format.

In addition, both unilex and preproce are now to be found in the lib directory and are common
to both the Unicon compiler and the parser. This will require updating the Makefile for the
Unicon compiler to copy the unilex.u and the preproce.u files to the lib directory when recompiled
for the Unicon compiler.

Signed-off-by: Bruce Rennie <brennie@dcsi.net.au>